Unify `theme()` functions in plugins and `resolveConfig()`
